+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/service/SyntaxService.java
@@ -53,6 +53,7 @@ public class SyntaxService {
             .on(NEWLINE)
             .withKeyValueSeparator(Splitter
                 .on(":")
+                .limit(2)
                 .trimResults()
                 .omitEmptyStrings())
             .split(prependSummaryKey);